правки по комментариям
